update dotValve with new dotClient pattern
